kernel: Add keyboard logger
We can finally type text on the keyboard and display it on the screen.

The scan codes is based on 'scan code set 1' which has 'US QWERTY'
layout. I tried with the Norwegian, but some of the characters is not part
of the standard ASCII character set.

The Norwegian keyboard layout was found at this site
<https://kbdlayout.info/KBDNO/scancodes>